+++ sum_price.c
@@ -2,7 +2,7 @@
 #include"class.h"
 
 //求单个订单的折后价
-void sumPrice(order* head, recipe* head_r, enum card c, int* d)
+void sumPrice(order* head, recipe* head_r, enum card c, double* d)
 {
 	head->sum_price = 0;
 	while (head_r != NULL)
